feat(inventory): sqm products + service materials drawn from stock (tasks 50, 51, 54)
تاسك 51 — المخزون بالمتر المربع:
كل كميات المخزون كانت integer فبيع «٠٫٥ م²» مستحيل. هجرة على النوع وحده تحوّل
stock_movements.qty و product_invoice_lines.qty و products.current_stock/min_stock_level
وسطور الجرد وكميات أوامر الشراء إلى DECIMAL(12,2) — لا صفّ واحد من stock_movements
يُمَسّ، فهو جدول إدراج فقط. و products.is_sqm يجعل نقطة البيع تطلب العرض والطول وعدد
القطع، والخادم يشتقّ الكمية = (العرض/100)×(الطول/100)×عدد القطع ويخزّنها مع المقاسين
على السطر — ما ترسله الواجهة لا يُعتمد. ومنتج القطعة لم يتغيّر.

تاسك 50 (+ شقّ المخزون من تاسك 54) — خامات الخدمة تُخصم من المخزون:
جدول branch_service_materials يربط خدمة الفرع بمنتجات مخزونه بكمية استهلاك لكل وحدة
محاسَب عليها في السطر (قطعة، أو م² لخدمة مسعّرة بالمتر). الخصم عند اعتماد الفاتورة لا
عند إنشائها آجلة — نفس لحظة العمولة والنقاط — والاسترجاع يكتب return_in مقابلاً يُقرأ
من الحركات المكتوبة لا من التعريف، فخدمةٌ تغيّرت خاماتها بعدُ تظلّ توازن إلى صفر.
تكلفة الخامات المحاسبية (تاسك 7) مستقلّة عنها تماماً ولا تتضاعف.

أعطال جانبية أصلحها الاختبار:
- StockMovementController::store كان يقرأ integer('qty') فيبتلع كسر المتر ويكتب حركة بصفر.
- BranchService::materials/userCommissions تحتاجان تسمية المفتاح صراحةً — الموديل Pivot
  و AsPivot::getForeignKey() يعيد خاصية الـpivot الفارغة.

25 اختباراً جديداً (SqmProductPosTest + ServiceMaterialsStockTest)، والسويت 1017 خضراء.

+++ app/Actions/Refund/CreateRefundAction.php
@@ -134,7 +134,7 @@ class CreateRefundAction
             $this->recordStockMovement->handle(
                 $line->product,
                 StockMovementTypeEnum::RETURN_IN,
-                (int) $line->qty,
+                (float) $line->qty,
                 [
                     'unit_cost' => $line->product->cost_price,
                     'reference_id' => $refund->id,
